clda-enhance/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
real-esrgan==0.3.0
torch>=2.0.0
basicsr>=1.4.2
